feat: 사용되고 있는 int를 Money로 변경

+++ src/main/java/lotto/view/ResultView.java
@@ -40,7 +40,7 @@ public class ResultView {
     private static String getFormatString(Map<LottoPrize, Integer> statisticsMap, LottoPrize lottoPrize) {
         return String.format(findMessageFormat(lottoPrize),
                              lottoPrize.getMatchCount(),
-                             lottoPrize.getPrizeMoney(),
+                             lottoPrize.getPrizeMoney().toInt(),
                              statisticsMap.get(lottoPrize));
     }
